+++ RELEASE_NOTES.md
@@ -4,6 +4,8 @@ v0.78.0
 
 - Operations have been consistently renamed to `store` and `restore`.
 - `restore` now returns an object containing both the BackupForwardSecrecyToken for decryption, and "secret data" to be used in the first `store` after restoration.
+- `SvrB` now has a `createNewBackupChain` method, allowing you to locally persist the backup "secret data" *before* the first store to SVR-B for a fresh install.
+- With the first two changes, the secret data argument to `restore` is now required. See doc comments for more details.
 
 ## Other changes